+++ src/main/kotlin/com/yapp/weekand/domain/user/dto/SearchUserListCondition.kt
@@ -1,0 +1,10 @@
+package com.yapp.weekand.domain.user.dto
+
+import com.yapp.weekand.api.generated.types.SearchUserSort
+
+data class SearchUserListCondition(
+	val nickNameOrGoalQuery: String?,
+	val jobs: List<String>?,
+	val interests: List<String>?,
+	val sort: SearchUserSort?
+)
